Network.AWS.WorkDocs

Description

The WorkDocs API is designed for the following use cases:

File Migration: File migration applications are supported for users who want to migrate their files from an on-premise or off-premise file system or service. Users can insert files into a user directory structure, as well as allow for basic metadata changes, such as modifications to the permissions of files.

Security: Support security applications are supported for users who have additional security needs, such as anti-virus or data loss prevention. The APIs, in conjunction with Amazon CloudTrail, allow these applications to detect when changes occur in Amazon WorkDocs, so the application can take the necessary actions and replace the target file. The application can also choose to email the user if the target file violates the policy.

eDiscoveryAnalytics: General administrative applications are supported, such as eDiscovery and analytics. These applications can choose to mimic andor record the actions in an Amazon WorkDocs site, in conjunction with Amazon CloudTrails, to replicate data for eDiscovery, backup, or analytical applications.

All Amazon WorkDocs APIs are Amazon authenticated, certificate-signed APIs. They not only require the use of the AWS SDK, but also allow for the exclusive use of IAM users and roles to help facilitate access, trust, and permission policies. By creating a role and allowing an IAM user to access the Amazon WorkDocs site, the IAM user gains full administrative visibility into the entire Amazon WorkDocs site (or as set in the IAM policy). This includes, but is not limited to, the ability to modify file permissions and upload any file to any user. This allows developers to perform the three use cases above, as well as give users the ability to grant access on a selective basis using the IAM model.

Service Configuration

workDocs :: Service

API version 2016-05-01 of the Amazon WorkDocs SDK configuration.

Errors

Error matchers are designed for use with the functions provided by Control.Exception.Lens. This allows catching (and rethrowing) service specific errors returned by WorkDocs.

CustomMetadataLimitExceededException

_CustomMetadataLimitExceededException :: AsError a => Getting (First ServiceError) a ServiceError

The limit has been reached on the number of custom properties for the specified resource.

EntityAlreadyExistsException

_EntityAlreadyExistsException :: AsError a => Getting (First ServiceError) a ServiceError

The resource already exists.

ResourceAlreadyCheckedOutException

_ResourceAlreadyCheckedOutException :: AsError a => Getting (First ServiceError) a ServiceError

The resource is already checked out.

ProhibitedStateException

_ProhibitedStateException :: AsError a => Getting (First ServiceError) a ServiceError

The specified document version is not in the INITIALIZED state.

TooManyLabelsException

_TooManyLabelsException :: AsError a => Getting (First ServiceError) a ServiceError

The limit has been reached on the number of labels for the specified resource.

InvalidArgumentException

_InvalidArgumentException :: AsError a => Getting (First ServiceError) a ServiceError

The pagination marker and/or limit fields are not valid.

UnauthorizedResourceAccessException

_UnauthorizedResourceAccessException :: AsError a => Getting (First ServiceError) a ServiceError

The caller does not have access to perform the action on the resource.

TooManySubscriptionsException

_TooManySubscriptionsException :: AsError a => Getting (First ServiceError) a ServiceError

You've reached the limit on the number of subscriptions for the WorkDocs instance.

FailedDependencyException

_FailedDependencyException :: AsError a => Getting (First ServiceError) a ServiceError

The AWS Directory Service cannot reach an on-premises instance. Or a dependency under the control of the organization is failing, such as a connected active directory.

DocumentLockedForCommentsException

_DocumentLockedForCommentsException :: AsError a => Getting (First ServiceError) a ServiceError

This exception is thrown when the document is locked for comments and user tries to create or delete a comment on that document.

EntityNotExistsException

_EntityNotExistsException :: AsError a => Getting (First ServiceError) a ServiceError

The resource does not exist.

DeactivatingLastSystemUserException

_DeactivatingLastSystemUserException :: AsError a => Getting (First ServiceError) a ServiceError

The last user in the organization is being deactivated.

IllegalUserStateException

_IllegalUserStateException :: AsError a => Getting (First ServiceError) a ServiceError

The user is undergoing transfer of ownership.

StorageLimitWillExceedException

_StorageLimitWillExceedException :: AsError a => Getting (First ServiceError) a ServiceError

The storage limit will be exceeded.

ConcurrentModificationException

_ConcurrentModificationException :: AsError a => Getting (First ServiceError) a ServiceError

The resource hierarchy is changing.

StorageLimitExceededException

_StorageLimitExceededException :: AsError a => Getting (First ServiceError) a ServiceError

The storage limit has been exceeded.

ServiceUnavailableException

_ServiceUnavailableException :: AsError a => Getting (First ServiceError) a ServiceError

One or more of the dependencies is unavailable.

InvalidOperationException

_InvalidOperationException :: AsError a => Getting (First ServiceError) a ServiceError

The operation is invalid.

UnauthorizedOperationException

_UnauthorizedOperationException :: AsError a => Getting (First ServiceError) a ServiceError

The operation is not permitted.

DraftUploadOutOfSyncException

_DraftUploadOutOfSyncException :: AsError a => Getting (First ServiceError) a ServiceError

This exception is thrown when a valid checkout ID is not presented on document version upload calls for a document that has been checked out from Web client.

LimitExceededException

_LimitExceededException :: AsError a => Getting (First ServiceError) a ServiceError

The maximum of 100,000 folders under the parent folder has been exceeded.

Waiters

Waiters poll by repeatedly sending a request until some remote success condition configured by the Wait specification is fulfilled. The Wait specification determines how many attempts should be made, in addition to delay and retry strategies.

Operations

Some AWS operations return results that are incomplete and require subsequent requests in order to obtain the entire result set. The process of sending subsequent requests to continue where a previous request left off is called pagination. For example, the ListObjects operation of Amazon S3 returns up to 1000 objects at a time, and you must send subsequent requests with the appropriate Marker in order to retrieve the next page of results.

Operations that have an AWSPager instance can transparently perform subsequent requests, correctly setting Markers and other request facets to iterate through the entire result set of a truncated API operation. Operations which support this have an additional note in the documentation.

Many operations have the ability to filter results on the server side. See the individual operation parameters for details.

DeleteLabels

module Network.AWS.WorkDocs.DeleteLabels

AbortDocumentVersionUpload

module Network.AWS.WorkDocs.AbortDocumentVersionUpload

GetDocumentPath

module Network.AWS.WorkDocs.GetDocumentPath

CreateComment

module Network.AWS.WorkDocs.CreateComment

DescribeUsers (Paginated)

module Network.AWS.WorkDocs.DescribeUsers

DeleteFolder

module Network.AWS.WorkDocs.DeleteFolder

UpdateFolder

module Network.AWS.WorkDocs.UpdateFolder

DeleteCustomMetadata

module Network.AWS.WorkDocs.DeleteCustomMetadata

DescribeResourcePermissions

module Network.AWS.WorkDocs.DescribeResourcePermissions

DeleteNotificationSubscription

module Network.AWS.WorkDocs.DeleteNotificationSubscription

CreateFolder

module Network.AWS.WorkDocs.CreateFolder

CreateNotificationSubscription

module Network.AWS.WorkDocs.CreateNotificationSubscription

CreateCustomMetadata

module Network.AWS.WorkDocs.CreateCustomMetadata

GetFolderPath

module Network.AWS.WorkDocs.GetFolderPath

DescribeComments

module Network.AWS.WorkDocs.DescribeComments

DeleteFolderContents

module Network.AWS.WorkDocs.DeleteFolderContents

RemoveAllResourcePermissions

module Network.AWS.WorkDocs.RemoveAllResourcePermissions

GetFolder

module Network.AWS.WorkDocs.GetFolder

DescribeNotificationSubscriptions

module Network.AWS.WorkDocs.DescribeNotificationSubscriptions

ActivateUser

module Network.AWS.WorkDocs.ActivateUser

DescribeDocumentVersions (Paginated)

module Network.AWS.WorkDocs.DescribeDocumentVersions

GetDocumentVersion

module Network.AWS.WorkDocs.GetDocumentVersion

DescribeActivities

module Network.AWS.WorkDocs.DescribeActivities

DescribeRootFolders

module Network.AWS.WorkDocs.DescribeRootFolders

GetCurrentUser

module Network.AWS.WorkDocs.GetCurrentUser

DeactivateUser

module Network.AWS.WorkDocs.DeactivateUser

GetDocument

module Network.AWS.WorkDocs.GetDocument

DescribeFolderContents (Paginated)

module Network.AWS.WorkDocs.DescribeFolderContents

CreateLabels

module Network.AWS.WorkDocs.CreateLabels

UpdateDocumentVersion

module Network.AWS.WorkDocs.UpdateDocumentVersion

RemoveResourcePermission

module Network.AWS.WorkDocs.RemoveResourcePermission

DeleteComment

module Network.AWS.WorkDocs.DeleteComment

InitiateDocumentVersionUpload

module Network.AWS.WorkDocs.InitiateDocumentVersionUpload

CreateUser

module Network.AWS.WorkDocs.CreateUser

UpdateUser

module Network.AWS.WorkDocs.UpdateUser

DeleteUser

module Network.AWS.WorkDocs.DeleteUser

AddResourcePermissions

module Network.AWS.WorkDocs.AddResourcePermissions

UpdateDocument

module Network.AWS.WorkDocs.UpdateDocument

DeleteDocument

module Network.AWS.WorkDocs.DeleteDocument

Types

ActivityType

data ActivityType

Constructors

DocumentAnnotationAdded

DocumentAnnotationDeleted

DocumentCheckedIn

DocumentCheckedOut

DocumentCommentAdded

DocumentCommentDeleted

DocumentMoved

DocumentRecycled

DocumentRenamed

DocumentRestored

DocumentReverted

DocumentSharePermissionChanged

DocumentShareableLinkCreated

DocumentShareableLinkPermissionChanged

DocumentShareableLinkRemoved

DocumentShared

DocumentUnshared

DocumentVersionDeleted

DocumentVersionUploaded

FolderCreated

FolderDeleted

FolderMoved

FolderRecycled

FolderRenamed

FolderRestored

FolderSharePermissionChanged

FolderShareableLinkCreated

FolderShareableLinkPermissionChanged

FolderShareableLinkRemoved

FolderShared

FolderUnshared

CommentStatusType

data CommentStatusType

Constructors

Deleted

Draft

Published

CommentVisibilityType

data CommentVisibilityType

Constructors

Private

Public

DocumentSourceType

data DocumentSourceType

Constructors

Original

WithComments

DocumentStatusType

data DocumentStatusType

Constructors

DSTActive

DSTInitialized

DocumentThumbnailType

data DocumentThumbnailType

Constructors

Large

Small

SmallHq

DocumentVersionStatus

data DocumentVersionStatus

Constructors

DVSActive

FolderContentType

data FolderContentType

Constructors

FCTAll

FCTDocument

FCTFolder

LocaleType

data LocaleType

Constructors

DE

Default

EN

ES

FR

JA

KO

PtBr

RU

ZhCn

ZhTw

OrderType

data OrderType

Constructors

Ascending

Descending

PrincipalType

data PrincipalType

Constructors

Anonymous

Group

Invite

Organization

User

ResourceSortType

data ResourceSortType

Constructors

Date

Name

ResourceStateType

data ResourceStateType

Constructors

RSTActive

RSTRecycled

RSTRecycling

RSTRestoring

ResourceType

data ResourceType

Constructors

Document

Folder

RolePermissionType

data RolePermissionType

Constructors

Direct

Inherited

RoleType

data RoleType

Constructors

Contributor

Coowner

Owner

Viewer

ShareStatusType

data ShareStatusType

Constructors

Failure

Success

StorageType

data StorageType

Constructors

Quota

Unlimited

SubscriptionProtocolType

data SubscriptionProtocolType

Constructors

HTTPS

SubscriptionType

data SubscriptionType

Constructors

STAll

UserFilterType

data UserFilterType

Constructors

ActivePending

All

UserSortType

data UserSortType

Constructors

FullName

StorageLimit

StorageUsed

UserName

UserStatus

UserStatusType

data UserStatusType

Constructors

Active

Inactive

Pending

UserType

data UserType

Constructors

UTAdmin

UTUser

Activity

data Activity

Describes the activity information.

See: activity smart constructor.

activity :: Activity

Creates a value of Activity with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

aResourceMetadata - The metadata of the resource involved in the user action.

aInitiator - The user who performed the action.

aParticipants - The list of users or groups impacted by this action. This is an optional field and is filled for the following sharing activities: DOCUMENT_SHARED, DOCUMENT_SHARED, DOCUMENT_UNSHARED, FOLDER_SHARED, FOLDER_UNSHARED.

aOriginalParent - The original parent of the resource. This is an optional field and is filled for move activities.

aType - The activity type.

aCommentMetadata - Metadata of the commenting activity. This is an optional field and is filled for commenting activities.

aTimeStamp - The timestamp when the action was performed.

aOrganizationId - The ID of the organization.

aResourceMetadata :: Lens' Activity (Maybe ResourceMetadata)

The metadata of the resource involved in the user action.

aInitiator :: Lens' Activity (Maybe UserMetadata)

The user who performed the action.

aParticipants :: Lens' Activity (Maybe Participants)

The list of users or groups impacted by this action. This is an optional field and is filled for the following sharing activities: DOCUMENT_SHARED, DOCUMENT_SHARED, DOCUMENT_UNSHARED, FOLDER_SHARED, FOLDER_UNSHARED.

aOriginalParent :: Lens' Activity (Maybe ResourceMetadata)

The original parent of the resource. This is an optional field and is filled for move activities.

aType :: Lens' Activity (Maybe ActivityType)

The activity type.

aCommentMetadata :: Lens' Activity (Maybe CommentMetadata)

Metadata of the commenting activity. This is an optional field and is filled for commenting activities.

aTimeStamp :: Lens' Activity (Maybe UTCTime)

The timestamp when the action was performed.

aOrganizationId :: Lens' Activity (Maybe Text)

The ID of the organization.

Comment

data Comment

Describes a comment.

See: comment smart constructor.

comment

Arguments

:: Text
cCommentId

-> Comment

Creates a value of Comment with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cStatus - The status of the comment.

cText - The text of the comment.

cVisibility - The visibility of the comment. Options are either PRIVATE, where the comment is visible only to the comment author and document owner and co-owners, or PUBLIC, where the comment is visible to document owners, co-owners, and contributors.

cThreadId - The ID of the root comment in the thread.

cContributor - The details of the user who made the comment.

cCreatedTimestamp - The time that the comment was created.

cRecipientId - If the comment is a reply to another user's comment, this field contains the user ID of the user being replied to.

cParentId - The ID of the parent comment.

cCommentId - The ID of the comment.

cStatus :: Lens' Comment (Maybe CommentStatusType)

The status of the comment.

cText :: Lens' Comment (Maybe Text)

The text of the comment.

cVisibility :: Lens' Comment (Maybe CommentVisibilityType)

The visibility of the comment. Options are either PRIVATE, where the comment is visible only to the comment author and document owner and co-owners, or PUBLIC, where the comment is visible to document owners, co-owners, and contributors.

cThreadId :: Lens' Comment (Maybe Text)

The ID of the root comment in the thread.

cContributor :: Lens' Comment (Maybe User)

The details of the user who made the comment.

cCreatedTimestamp :: Lens' Comment (Maybe UTCTime)

The time that the comment was created.

cRecipientId :: Lens' Comment (Maybe Text)

If the comment is a reply to another user's comment, this field contains the user ID of the user being replied to.

cParentId :: Lens' Comment (Maybe Text)

The ID of the parent comment.

cCommentId :: Lens' Comment Text

The ID of the comment.

CommentMetadata

data CommentMetadata

Describes the metadata of a comment.

See: commentMetadata smart constructor.

commentMetadata :: CommentMetadata

Creates a value of CommentMetadata with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cmCommentStatus - Undocumented member.

cmContributor - The user who made the comment.

cmCommentId - The ID of the comment.

cmCreatedTimestamp - Undocumented member.

cmRecipientId - The ID of the user being replied to.

cmCommentStatus :: Lens' CommentMetadata (Maybe CommentStatusType)

Undocumented member.

cmContributor :: Lens' CommentMetadata (Maybe User)

The user who made the comment.

cmCommentId :: Lens' CommentMetadata (Maybe Text)

The ID of the comment.

cmCreatedTimestamp :: Lens' CommentMetadata (Maybe UTCTime)

Undocumented member.

cmRecipientId :: Lens' CommentMetadata (Maybe Text)

The ID of the user being replied to.

DocumentMetadata

data DocumentMetadata

Describes the document.

See: documentMetadata smart constructor.

documentMetadata :: DocumentMetadata

Creates a value of DocumentMetadata with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dmLatestVersionMetadata - The latest version of the document.

dmParentFolderId - The ID of the parent folder.

dmModifiedTimestamp - The time when the document was updated.

dmId - The ID of the document.

dmLabels - List of labels on the document.

dmResourceState - The resource state.

dmCreatedTimestamp - The time when the document was created.

dmCreatorId - The ID of the creator.

dmLatestVersionMetadata :: Lens' DocumentMetadata (Maybe DocumentVersionMetadata)

The latest version of the document.

dmParentFolderId :: Lens' DocumentMetadata (Maybe Text)

The ID of the parent folder.

dmModifiedTimestamp :: Lens' DocumentMetadata (Maybe UTCTime)

The time when the document was updated.

dmId :: Lens' DocumentMetadata (Maybe Text)

The ID of the document.

dmLabels :: Lens' DocumentMetadata [Text]

List of labels on the document.

dmResourceState :: Lens' DocumentMetadata (Maybe ResourceStateType)

The resource state.

dmCreatedTimestamp :: Lens' DocumentMetadata (Maybe UTCTime)

The time when the document was created.

dmCreatorId :: Lens' DocumentMetadata (Maybe Text)

The ID of the creator.

DocumentVersionMetadata

data DocumentVersionMetadata

Describes a version of a document.

See: documentVersionMetadata smart constructor.

documentVersionMetadata :: DocumentVersionMetadata

Creates a value of DocumentVersionMetadata with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dvmThumbnail - The thumbnail of the document.

dvmStatus - The status of the document.

dvmSignature - The signature of the document.

dvmContentCreatedTimestamp - The time stamp when the content of the document was originally created.

dvmSize - The size of the document, in bytes.

dvmName - The name of the version.

dvmModifiedTimestamp - The time stamp when the document was last uploaded.

dvmSource - The source of the document.

dvmId - The ID of the version.

dvmCreatedTimestamp - The time stamp when the document was first uploaded.

dvmContentModifiedTimestamp - The time stamp when the content of the document was modified.

dvmCreatorId - The ID of the creator.

dvmContentType - The content type of the document.

dvmThumbnail :: Lens' DocumentVersionMetadata (HashMap DocumentThumbnailType Text)

The thumbnail of the document.

dvmStatus :: Lens' DocumentVersionMetadata (Maybe DocumentStatusType)

The status of the document.

dvmSignature :: Lens' DocumentVersionMetadata (Maybe Text)

The signature of the document.

dvmContentCreatedTimestamp :: Lens' DocumentVersionMetadata (Maybe UTCTime)

The time stamp when the content of the document was originally created.

dvmSize :: Lens' DocumentVersionMetadata (Maybe Integer)

The size of the document, in bytes.

dvmName :: Lens' DocumentVersionMetadata (Maybe Text)

The name of the version.

dvmModifiedTimestamp :: Lens' DocumentVersionMetadata (Maybe UTCTime)

The time stamp when the document was last uploaded.

dvmSource :: Lens' DocumentVersionMetadata (HashMap DocumentSourceType Text)

The source of the document.

dvmId :: Lens' DocumentVersionMetadata (Maybe Text)

The ID of the version.

dvmCreatedTimestamp :: Lens' DocumentVersionMetadata (Maybe UTCTime)

The time stamp when the document was first uploaded.

dvmContentModifiedTimestamp :: Lens' DocumentVersionMetadata (Maybe UTCTime)

The time stamp when the content of the document was modified.

dvmCreatorId :: Lens' DocumentVersionMetadata (Maybe Text)

The ID of the creator.

dvmContentType :: Lens' DocumentVersionMetadata (Maybe Text)

The content type of the document.

FolderMetadata

data FolderMetadata

Describes a folder.

See: folderMetadata smart constructor.

folderMetadata :: FolderMetadata

Creates a value of FolderMetadata with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

fmSignature - The unique identifier created from the subfolders and documents of the folder.

fmParentFolderId - The ID of the parent folder.

fmSize - The size of the folder metadata.

fmLatestVersionSize - The size of the latest version of the folder metadata.

fmName - The name of the folder.

fmModifiedTimestamp - The time when the folder was updated.

fmId - The ID of the folder.

fmLabels - List of labels on the folder.

fmResourceState - The resource state of the folder.

fmCreatedTimestamp - The time when the folder was created.

fmCreatorId - The ID of the creator.

fmSignature :: Lens' FolderMetadata (Maybe Text)

The unique identifier created from the subfolders and documents of the folder.

fmParentFolderId :: Lens' FolderMetadata (Maybe Text)

The ID of the parent folder.

fmSize :: Lens' FolderMetadata (Maybe Integer)

The size of the folder metadata.

fmLatestVersionSize :: Lens' FolderMetadata (Maybe Integer)

The size of the latest version of the folder metadata.

fmName :: Lens' FolderMetadata (Maybe Text)

The name of the folder.

fmModifiedTimestamp :: Lens' FolderMetadata (Maybe UTCTime)

The time when the folder was updated.

fmId :: Lens' FolderMetadata (Maybe Text)

The ID of the folder.

fmLabels :: Lens' FolderMetadata [Text]

List of labels on the folder.

fmResourceState :: Lens' FolderMetadata (Maybe ResourceStateType)

The resource state of the folder.

fmCreatedTimestamp :: Lens' FolderMetadata (Maybe UTCTime)

The time when the folder was created.

fmCreatorId :: Lens' FolderMetadata (Maybe Text)

The ID of the creator.

GroupMetadata

data GroupMetadata

Describes the metadata of a user group.

See: groupMetadata smart constructor.

groupMetadata :: GroupMetadata

Creates a value of GroupMetadata with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gmName - The name of the group.

gmId - The ID of the user group.

gmName :: Lens' GroupMetadata (Maybe Text)

The name of the group.

gmId :: Lens' GroupMetadata (Maybe Text)

The ID of the user group.

Participants

data Participants

Describes the users and/or user groups.

See: participants smart constructor.

participants :: Participants

Creates a value of Participants with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

pGroups - The list of user groups.

pUsers - The list of users.

pGroups :: Lens' Participants [GroupMetadata]

The list of user groups.

pUsers :: Lens' Participants [UserMetadata]

The list of users.

PermissionInfo

data PermissionInfo

Describes the permissions.

See: permissionInfo smart constructor.

permissionInfo :: PermissionInfo

Creates a value of PermissionInfo with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

piRole - The role of the user.

piType - The type of permissions.

piRole :: Lens' PermissionInfo (Maybe RoleType)

The role of the user.

piType :: Lens' PermissionInfo (Maybe RolePermissionType)

The type of permissions.

Principal

data Principal

Describes a resource.

See: principal smart constructor.

principal :: Principal

Creates a value of Principal with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

pRoles - The permission information for the resource.

pId - The ID of the resource.

pType - The type of resource.

pRoles :: Lens' Principal [PermissionInfo]

The permission information for the resource.

pId :: Lens' Principal (Maybe Text)

The ID of the resource.

pType :: Lens' Principal (Maybe PrincipalType)

The type of resource.

ResourceMetadata

data ResourceMetadata

Describes the metadata of a resource.

See: resourceMetadata smart constructor.

resourceMetadata :: ResourceMetadata

Creates a value of ResourceMetadata with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

rmVersionId - The version ID of the resource. This is an optional field and is filled for action on document version.

rmOwner - The owner of the resource.

rmName - The name of the resource.

rmId - The ID of the resource.

rmType - The type of resource.

rmOriginalName - The original name of the resource prior to a rename operation.

rmParentId - The parent ID of the resource before a rename operation.

rmVersionId :: Lens' ResourceMetadata (Maybe Text)

The version ID of the resource. This is an optional field and is filled for action on document version.

rmOwner :: Lens' ResourceMetadata (Maybe UserMetadata)

The owner of the resource.

rmName :: Lens' ResourceMetadata (Maybe Text)

The name of the resource.

rmId :: Lens' ResourceMetadata (Maybe Text)

The ID of the resource.

rmType :: Lens' ResourceMetadata (Maybe ResourceType)

The type of resource.

rmOriginalName :: Lens' ResourceMetadata (Maybe Text)

The original name of the resource prior to a rename operation.

rmParentId :: Lens' ResourceMetadata (Maybe Text)

The parent ID of the resource before a rename operation.

ResourcePath

data ResourcePath

Describes the path information of a resource.

See: resourcePath smart constructor.

resourcePath :: ResourcePath

Creates a value of ResourcePath with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

rpComponents - The components of the resource path.

rpComponents :: Lens' ResourcePath [ResourcePathComponent]

The components of the resource path.

ResourcePathComponent

data ResourcePathComponent

Describes the resource path.

See: resourcePathComponent smart constructor.

resourcePathComponent :: ResourcePathComponent

Creates a value of ResourcePathComponent with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

rpcName - The name of the resource path.

rpcId - The ID of the resource path.

rpcName :: Lens' ResourcePathComponent (Maybe Text)

The name of the resource path.

rpcId :: Lens' ResourcePathComponent (Maybe Text)

The ID of the resource path.

SharePrincipal

data SharePrincipal

Describes the recipient type and ID, if available.

See: sharePrincipal smart constructor.

sharePrincipal

Arguments

:: Text
spId

-> PrincipalType
spType

-> RoleType
spRole

-> SharePrincipal

Creates a value of SharePrincipal with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

spId - The ID of the recipient.

spType - The type of the recipient.

spRole - The role of the recipient.

spId :: Lens' SharePrincipal Text

The ID of the recipient.

spType :: Lens' SharePrincipal PrincipalType

The type of the recipient.

spRole :: Lens' SharePrincipal RoleType

The role of the recipient.

ShareResult

data ShareResult

Describes the share results of a resource.

See: shareResult smart constructor.

shareResult :: ShareResult

Creates a value of ShareResult with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

srStatus - The status.

srPrincipalId - The ID of the principal.

srRole - The role.

srStatusMessage - The status message.

srShareId - The ID of the resource that was shared.

srStatus :: Lens' ShareResult (Maybe ShareStatusType)

The status.

srPrincipalId :: Lens' ShareResult (Maybe Text)

The ID of the principal.

srRole :: Lens' ShareResult (Maybe RoleType)

The role.

srStatusMessage :: Lens' ShareResult (Maybe Text)

The status message.

srShareId :: Lens' ShareResult (Maybe Text)

The ID of the resource that was shared.

StorageRuleType

data StorageRuleType

Describes the storage for a user.

See: storageRuleType smart constructor.

storageRuleType :: StorageRuleType

Creates a value of StorageRuleType with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

srtStorageAllocatedInBytes - The amount of storage allocated, in bytes.

srtStorageType - The type of storage.

srtStorageAllocatedInBytes :: Lens' StorageRuleType (Maybe Natural)

The amount of storage allocated, in bytes.

srtStorageType :: Lens' StorageRuleType (Maybe StorageType)

The type of storage.

Subscription

data Subscription

Describes a subscription.

See: subscription smart constructor.

subscription :: Subscription

Creates a value of Subscription with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

sProtocol - The protocol of the subscription.

sEndPoint - The endpoint of the subscription.

sSubscriptionId - The ID of the subscription.

sProtocol :: Lens' Subscription (Maybe SubscriptionProtocolType)

The protocol of the subscription.

sEndPoint :: Lens' Subscription (Maybe Text)

The endpoint of the subscription.

sSubscriptionId :: Lens' Subscription (Maybe Text)

The ID of the subscription.

UploadMetadata

data UploadMetadata

Describes the upload.

See: uploadMetadata smart constructor.

uploadMetadata :: UploadMetadata

Creates a value of UploadMetadata with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

umUploadURL - The URL of the upload.

umSignedHeaders - The signed headers.

umUploadURL :: Lens' UploadMetadata (Maybe Text)

The URL of the upload.

umSignedHeaders :: Lens' UploadMetadata (HashMap Text Text)

The signed headers.

User

data User

Describes a user.

See: user smart constructor.

user :: User

Creates a value of User with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

uGivenName - The given name of the user.

uStatus - The status of the user.

uLocale - The locale of the user.

uUsername - The login name of the user.

uStorage - The storage for the user.

uModifiedTimestamp - The time when the user was modified.

uEmailAddress - The email address of the user.

uId - The ID of the user.

uRootFolderId - The ID of the root folder.

uType - The type of user.

uSurname - The surname of the user.

uTimeZoneId - The time zone ID of the user.

uCreatedTimestamp - The time when the user was created.

uOrganizationId - The ID of the organization.

uRecycleBinFolderId - The ID of the recycle bin folder.

uGivenName :: Lens' User (Maybe Text)

The given name of the user.

uStatus :: Lens' User (Maybe UserStatusType)

The status of the user.

uLocale :: Lens' User (Maybe LocaleType)

The locale of the user.

uUsername :: Lens' User (Maybe Text)

The login name of the user.

uStorage :: Lens' User (Maybe UserStorageMetadata)

The storage for the user.

uModifiedTimestamp :: Lens' User (Maybe UTCTime)

The time when the user was modified.

uEmailAddress :: Lens' User (Maybe Text)

The email address of the user.

uId :: Lens' User (Maybe Text)

The ID of the user.

uRootFolderId :: Lens' User (Maybe Text)

The ID of the root folder.

uType :: Lens' User (Maybe UserType)

The type of user.

uSurname :: Lens' User (Maybe Text)

The surname of the user.

uTimeZoneId :: Lens' User (Maybe Text)

The time zone ID of the user.

uCreatedTimestamp :: Lens' User (Maybe UTCTime)

The time when the user was created.

uOrganizationId :: Lens' User (Maybe Text)

The ID of the organization.

uRecycleBinFolderId :: Lens' User (Maybe Text)

The ID of the recycle bin folder.

UserMetadata

data UserMetadata

Describes the metadata of the user.

See: userMetadata smart constructor.

userMetadata :: UserMetadata

Creates a value of UserMetadata with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

umGivenName - The given name of the user before a rename operation.

umUsername - The username of the user.

umEmailAddress - The email address of the user.

umId - The ID of the user.

umSurname - The surname of the user.

umGivenName :: Lens' UserMetadata (Maybe Text)

The given name of the user before a rename operation.

umUsername :: Lens' UserMetadata (Maybe Text)

The username of the user.

umEmailAddress :: Lens' UserMetadata (Maybe Text)

The email address of the user.

umId :: Lens' UserMetadata (Maybe Text)

The ID of the user.

umSurname :: Lens' UserMetadata (Maybe Text)

The surname of the user.

UserStorageMetadata

data UserStorageMetadata

Describes the storage for a user.

See: userStorageMetadata smart constructor.

userStorageMetadata :: UserStorageMetadata

Creates a value of UserStorageMetadata with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

usmStorageUtilizedInBytes - The amount of storage utilized, in bytes.

usmStorageRule - The storage for a user.

usmStorageUtilizedInBytes :: Lens' UserStorageMetadata (Maybe Integer)

The amount of storage utilized, in bytes.

usmStorageRule :: Lens' UserStorageMetadata (Maybe StorageRuleType)

The storage for a user.