Network.AWS.WorkDocs.ActivateUser

Description

Activates the specified user. Only active users can access Amazon WorkDocs.

Creating a Request

activateUser

Arguments

:: Text
auUserId

-> ActivateUser

Creates a value of ActivateUser with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

auAuthenticationToken - Amazon WorkDocs authentication token. This field should not be set when using administrative API actions, as in accessing the API using AWS credentials.

auUserId - The ID of the user.

data ActivateUser

See: activateUser smart constructor.

Request Lenses

auAuthenticationToken :: Lens' ActivateUser (Maybe Text)

Amazon WorkDocs authentication token. This field should not be set when using administrative API actions, as in accessing the API using AWS credentials.

auUserId :: Lens' ActivateUser Text

The ID of the user.

Destructuring the Response

activateUserResponse

Arguments

:: Int
aursResponseStatus

-> ActivateUserResponse

Creates a value of ActivateUserResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

aursUser - The user information.

aursResponseStatus - -- | The response status code.

data ActivateUserResponse

See: activateUserResponse smart constructor.

Response Lenses

aursUser :: Lens' ActivateUserResponse (Maybe User)

The user information.

aursResponseStatus :: Lens' ActivateUserResponse Int

- | The response status code.