Network.AWS.WorkDocs.CreateCustomMetadata

Description

Adds one or more custom properties to the specified resource (a folder, document, or version).

Creating a Request

createCustomMetadata

Arguments

:: Text
ccmResourceId

-> CreateCustomMetadata

Creates a value of CreateCustomMetadata with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ccmVersionId - The ID of the version, if the custom metadata is being added to a document version.

ccmAuthenticationToken - Amazon WorkDocs authentication token. This field should not be set when using administrative API actions, as in accessing the API using AWS credentials.

ccmResourceId - The ID of the resource.

ccmCustomMetadata - Custom metadata in the form of name-value pairs.

data CreateCustomMetadata

See: createCustomMetadata smart constructor.

Request Lenses

ccmVersionId :: Lens' CreateCustomMetadata (Maybe Text)

The ID of the version, if the custom metadata is being added to a document version.

ccmAuthenticationToken :: Lens' CreateCustomMetadata (Maybe Text)

Amazon WorkDocs authentication token. This field should not be set when using administrative API actions, as in accessing the API using AWS credentials.

ccmResourceId :: Lens' CreateCustomMetadata Text

The ID of the resource.

ccmCustomMetadata :: Lens' CreateCustomMetadata (HashMap Text Text)

Custom metadata in the form of name-value pairs.

Destructuring the Response

createCustomMetadataResponse

Arguments

:: Int
ccmrsResponseStatus

-> CreateCustomMetadataResponse

Creates a value of CreateCustomMetadataResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ccmrsResponseStatus - -- | The response status code.

data CreateCustomMetadataResponse

See: createCustomMetadataResponse smart constructor.

Response Lenses

ccmrsResponseStatus :: Lens' CreateCustomMetadataResponse Int

- | The response status code.