Network.AWS.WorkDocs.CreateNotificationSubscription

Description

Configure WorkDocs to use Amazon SNS notifications.

The endpoint receives a confirmation message, and must confirm the subscription. For more information, see Confirm the Subscription in the Amazon Simple Notification Service Developer Guide .

Creating a Request

createNotificationSubscription

Arguments

:: Text
cnsOrganizationId

-> Text
cnsEndpoint

-> SubscriptionProtocolType
cnsProtocol

-> SubscriptionType
cnsSubscriptionType

-> CreateNotificationSubscription

Creates a value of CreateNotificationSubscription with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cnsOrganizationId - The ID of the organization.

cnsEndpoint - The endpoint to receive the notifications. If the protocol is HTTPS, the endpoint is a URL that begins with "https://".

cnsProtocol - The protocol to use. The supported value is https, which delivers JSON-encoded messasges using HTTPS POST.

cnsSubscriptionType - The notification type.

data CreateNotificationSubscription

See: createNotificationSubscription smart constructor.

Request Lenses

cnsOrganizationId :: Lens' CreateNotificationSubscription Text

The ID of the organization.

cnsEndpoint :: Lens' CreateNotificationSubscription Text

The endpoint to receive the notifications. If the protocol is HTTPS, the endpoint is a URL that begins with "https://".

cnsProtocol :: Lens' CreateNotificationSubscription SubscriptionProtocolType

The protocol to use. The supported value is https, which delivers JSON-encoded messasges using HTTPS POST.

cnsSubscriptionType :: Lens' CreateNotificationSubscription SubscriptionType

The notification type.

Destructuring the Response

createNotificationSubscriptionResponse

Arguments

:: Int
cnsrsResponseStatus

-> CreateNotificationSubscriptionResponse

Creates a value of CreateNotificationSubscriptionResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cnsrsSubscription - The subscription.

cnsrsResponseStatus - -- | The response status code.

data CreateNotificationSubscriptionResponse

See: createNotificationSubscriptionResponse smart constructor.

Response Lenses

cnsrsSubscription :: Lens' CreateNotificationSubscriptionResponse (Maybe Subscription)

The subscription.

cnsrsResponseStatus :: Lens' CreateNotificationSubscriptionResponse Int

- | The response status code.