Network.AWS.WorkDocs.CreateUser

Description

Creates a user in a Simple AD or Microsoft AD directory. The status of a newly created user is ACTIVE. New users can access Amazon WorkDocs.

Creating a Request

createUser

Arguments

:: Text
cuUsername

-> Text
cuGivenName

-> Text
cuSurname

-> Text
cuPassword

-> CreateUser

Creates a value of CreateUser with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cuAuthenticationToken - Amazon WorkDocs authentication token. This field should not be set when using administrative API actions, as in accessing the API using AWS credentials.

cuStorageRule - The amount of storage for the user.

cuEmailAddress - The email address of the user.

cuTimeZoneId - The time zone ID of the user.

cuOrganizationId - The ID of the organization.

cuUsername - The login name of the user.

cuGivenName - The given name of the user.

cuSurname - The surname of the user.

cuPassword - The password of the user.

data CreateUser

See: createUser smart constructor.

Request Lenses

cuAuthenticationToken :: Lens' CreateUser (Maybe Text)

Amazon WorkDocs authentication token. This field should not be set when using administrative API actions, as in accessing the API using AWS credentials.

cuStorageRule :: Lens' CreateUser (Maybe StorageRuleType)

The amount of storage for the user.

cuEmailAddress :: Lens' CreateUser (Maybe Text)

The email address of the user.

cuTimeZoneId :: Lens' CreateUser (Maybe Text)

The time zone ID of the user.

cuOrganizationId :: Lens' CreateUser (Maybe Text)

The ID of the organization.

cuUsername :: Lens' CreateUser Text

The login name of the user.

cuGivenName :: Lens' CreateUser Text

The given name of the user.

cuSurname :: Lens' CreateUser Text

The surname of the user.

cuPassword :: Lens' CreateUser Text

The password of the user.

Destructuring the Response

createUserResponse

Arguments

:: Int
cursResponseStatus

-> CreateUserResponse

Creates a value of CreateUserResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cursUser - The user information.

cursResponseStatus - -- | The response status code.

data CreateUserResponse

See: createUserResponse smart constructor.

Response Lenses

cursUser :: Lens' CreateUserResponse (Maybe User)

The user information.

cursResponseStatus :: Lens' CreateUserResponse Int

- | The response status code.