Network.AWS.WorkDocs.DeactivateUser

Description

Deactivates the specified user, which revokes the user's access to Amazon WorkDocs.

Creating a Request

deactivateUser

Arguments

:: Text
dUserId

-> DeactivateUser

Creates a value of DeactivateUser with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dAuthenticationToken - Amazon WorkDocs authentication token. This field should not be set when using administrative API actions, as in accessing the API using AWS credentials.

dUserId - The ID of the user.

data DeactivateUser

See: deactivateUser smart constructor.

Request Lenses

dAuthenticationToken :: Lens' DeactivateUser (Maybe Text)

Amazon WorkDocs authentication token. This field should not be set when using administrative API actions, as in accessing the API using AWS credentials.

dUserId :: Lens' DeactivateUser Text

The ID of the user.

Destructuring the Response

deactivateUserResponse :: DeactivateUserResponse

Creates a value of DeactivateUserResponse with the minimum fields required to make a request.

data DeactivateUserResponse

See: deactivateUserResponse smart constructor.