Network.AWS.WorkDocs.DeleteFolder

Description

Permanently deletes the specified folder and its contents.

Creating a Request

deleteFolder

Arguments

:: Text
dfFolderId

-> DeleteFolder

Creates a value of DeleteFolder with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dfAuthenticationToken - Amazon WorkDocs authentication token. This field should not be set when using administrative API actions, as in accessing the API using AWS credentials.

dfFolderId - The ID of the folder.

data DeleteFolder

See: deleteFolder smart constructor.

Request Lenses

dfAuthenticationToken :: Lens' DeleteFolder (Maybe Text)

Amazon WorkDocs authentication token. This field should not be set when using administrative API actions, as in accessing the API using AWS credentials.

dfFolderId :: Lens' DeleteFolder Text

The ID of the folder.

Destructuring the Response

deleteFolderResponse :: DeleteFolderResponse

Creates a value of DeleteFolderResponse with the minimum fields required to make a request.

data DeleteFolderResponse

See: deleteFolderResponse smart constructor.