Network.AWS.WorkDocs.DeleteFolderContents

Description

Deletes the contents of the specified folder.

Creating a Request

deleteFolderContents

Arguments

:: Text
dfcFolderId

-> DeleteFolderContents

Creates a value of DeleteFolderContents with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dfcAuthenticationToken - Amazon WorkDocs authentication token. This field should not be set when using administrative API actions, as in accessing the API using AWS credentials.

dfcFolderId - The ID of the folder.

data DeleteFolderContents

See: deleteFolderContents smart constructor.

Request Lenses

dfcAuthenticationToken :: Lens' DeleteFolderContents (Maybe Text)

Amazon WorkDocs authentication token. This field should not be set when using administrative API actions, as in accessing the API using AWS credentials.

dfcFolderId :: Lens' DeleteFolderContents Text

The ID of the folder.

Destructuring the Response

deleteFolderContentsResponse :: DeleteFolderContentsResponse

Creates a value of DeleteFolderContentsResponse with the minimum fields required to make a request.

data DeleteFolderContentsResponse

See: deleteFolderContentsResponse smart constructor.