Network.AWS.WorkDocs.DeleteLabels

Description

Deletes the specified list of labels from a resource.

Creating a Request

deleteLabels

Arguments

:: Text
dlResourceId

-> DeleteLabels

Creates a value of DeleteLabels with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dlDeleteAll - Flag to request removal of all labels from the specified resource.

dlAuthenticationToken - Amazon WorkDocs authentication token. This field should not be set when using administrative API actions, as in accessing the API using AWS credentials.

dlLabels - List of labels to delete from the resource.

dlResourceId - The ID of the resource.

data DeleteLabels

See: deleteLabels smart constructor.

Request Lenses

dlDeleteAll :: Lens' DeleteLabels (Maybe Bool)

Flag to request removal of all labels from the specified resource.

dlAuthenticationToken :: Lens' DeleteLabels (Maybe Text)

Amazon WorkDocs authentication token. This field should not be set when using administrative API actions, as in accessing the API using AWS credentials.

dlLabels :: Lens' DeleteLabels [Text]

List of labels to delete from the resource.

dlResourceId :: Lens' DeleteLabels Text

The ID of the resource.

Destructuring the Response

deleteLabelsResponse

Arguments

:: Int
dlrsResponseStatus

-> DeleteLabelsResponse

Creates a value of DeleteLabelsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dlrsResponseStatus - -- | The response status code.

data DeleteLabelsResponse

See: deleteLabelsResponse smart constructor.

Response Lenses

dlrsResponseStatus :: Lens' DeleteLabelsResponse Int

- | The response status code.