Network.AWS.WorkDocs.DeleteNotificationSubscription

Description

Deletes the specified subscription from the specified organization.

Creating a Request

deleteNotificationSubscription

Arguments

:: Text
dnsSubscriptionId

-> Text
dnsOrganizationId

-> DeleteNotificationSubscription

Creates a value of DeleteNotificationSubscription with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dnsSubscriptionId - The ID of the subscription.

dnsOrganizationId - The ID of the organization.

data DeleteNotificationSubscription

See: deleteNotificationSubscription smart constructor.

Request Lenses

dnsSubscriptionId :: Lens' DeleteNotificationSubscription Text

The ID of the subscription.

dnsOrganizationId :: Lens' DeleteNotificationSubscription Text

The ID of the organization.

Destructuring the Response

deleteNotificationSubscriptionResponse :: DeleteNotificationSubscriptionResponse

Creates a value of DeleteNotificationSubscriptionResponse with the minimum fields required to make a request.

data DeleteNotificationSubscriptionResponse

See: deleteNotificationSubscriptionResponse smart constructor.