Network.AWS.WorkDocs.DescribeActivities

Description

Describes the user activities in a specified time period.

Creating a Request

describeActivities :: DescribeActivities

Creates a value of DescribeActivities with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

daStartTime - The timestamp that determines the starting time of the activities; the response includes the activities performed after the specified timestamp.

daAuthenticationToken - Amazon WorkDocs authentication token. This field should not be set when using administrative API actions, as in accessing the API using AWS credentials.

daUserId - The ID of the user who performed the action. The response includes activities pertaining to this user. This is an optional parameter and is only applicable for administrative API (SigV4) requests.

daMarker - The marker for the next set of results. (You received this marker from a previous call.)

daEndTime - The timestamp that determines the end time of the activities; the response includes the activities performed before the specified timestamp.

daLimit - The maximum number of items to return.

daOrganizationId - The ID of the organization. This is a mandatory parameter when using administrative API (SigV4) requests.

data DescribeActivities

See: describeActivities smart constructor.

Request Lenses

daStartTime :: Lens' DescribeActivities (Maybe UTCTime)

The timestamp that determines the starting time of the activities; the response includes the activities performed after the specified timestamp.

daAuthenticationToken :: Lens' DescribeActivities (Maybe Text)

Amazon WorkDocs authentication token. This field should not be set when using administrative API actions, as in accessing the API using AWS credentials.

daUserId :: Lens' DescribeActivities (Maybe Text)

The ID of the user who performed the action. The response includes activities pertaining to this user. This is an optional parameter and is only applicable for administrative API (SigV4) requests.

daMarker :: Lens' DescribeActivities (Maybe Text)

The marker for the next set of results. (You received this marker from a previous call.)

daEndTime :: Lens' DescribeActivities (Maybe UTCTime)

The timestamp that determines the end time of the activities; the response includes the activities performed before the specified timestamp.

daLimit :: Lens' DescribeActivities (Maybe Natural)

The maximum number of items to return.

daOrganizationId :: Lens' DescribeActivities (Maybe Text)

The ID of the organization. This is a mandatory parameter when using administrative API (SigV4) requests.

Destructuring the Response

describeActivitiesResponse

Arguments

:: Int
darsResponseStatus

-> DescribeActivitiesResponse

Creates a value of DescribeActivitiesResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

darsUserActivities - The list of activities for the specified user and time period.

darsMarker - The marker for the next set of results.

darsResponseStatus - -- | The response status code.

data DescribeActivitiesResponse

See: describeActivitiesResponse smart constructor.

Response Lenses

darsUserActivities :: Lens' DescribeActivitiesResponse [Activity]

The list of activities for the specified user and time period.

darsMarker :: Lens' DescribeActivitiesResponse (Maybe Text)

The marker for the next set of results.

darsResponseStatus :: Lens' DescribeActivitiesResponse Int

- | The response status code.