Network.AWS.WorkDocs.DescribeDocumentVersions

Description

Retrieves the document versions for the specified document.

By default, only active versions are returned.

This operation returns paginated results.

Creating a Request

describeDocumentVersions

Arguments

:: Text
ddvDocumentId

-> DescribeDocumentVersions

Creates a value of DescribeDocumentVersions with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ddvInclude - A comma-separated list of values. Specify INITIALIZED to include incomplete versions.

ddvAuthenticationToken - Amazon WorkDocs authentication token. This field should not be set when using administrative API actions, as in accessing the API using AWS credentials.

ddvMarker - The marker for the next set of results. (You received this marker from a previous call.)

ddvLimit - The maximum number of versions to return with this call.

ddvFields - Specify SOURCE to include initialized versions and a URL for the source document.

ddvDocumentId - The ID of the document.

data DescribeDocumentVersions

See: describeDocumentVersions smart constructor.

Request Lenses

ddvInclude :: Lens' DescribeDocumentVersions (Maybe Text)

A comma-separated list of values. Specify INITIALIZED to include incomplete versions.

ddvAuthenticationToken :: Lens' DescribeDocumentVersions (Maybe Text)

Amazon WorkDocs authentication token. This field should not be set when using administrative API actions, as in accessing the API using AWS credentials.

ddvMarker :: Lens' DescribeDocumentVersions (Maybe Text)

The marker for the next set of results. (You received this marker from a previous call.)

ddvLimit :: Lens' DescribeDocumentVersions (Maybe Natural)

The maximum number of versions to return with this call.

ddvFields :: Lens' DescribeDocumentVersions (Maybe Text)

Specify SOURCE to include initialized versions and a URL for the source document.

ddvDocumentId :: Lens' DescribeDocumentVersions Text

The ID of the document.

Destructuring the Response

describeDocumentVersionsResponse

Arguments

:: Int
ddvrsResponseStatus

-> DescribeDocumentVersionsResponse

Creates a value of DescribeDocumentVersionsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ddvrsDocumentVersions - The document versions.

ddvrsMarker - The marker to use when requesting the next set of results. If there are no additional results, the string is empty.

ddvrsResponseStatus - -- | The response status code.

data DescribeDocumentVersionsResponse

See: describeDocumentVersionsResponse smart constructor.

Response Lenses

ddvrsDocumentVersions :: Lens' DescribeDocumentVersionsResponse [DocumentVersionMetadata]

The document versions.

ddvrsMarker :: Lens' DescribeDocumentVersionsResponse (Maybe Text)

The marker to use when requesting the next set of results. If there are no additional results, the string is empty.

ddvrsResponseStatus :: Lens' DescribeDocumentVersionsResponse Int

- | The response status code.