Network.AWS.WorkDocs.DescribeFolderContents

Description

Describes the contents of the specified folder, including its documents and subfolders.

By default, Amazon WorkDocs returns the first 100 active document and folder metadata items. If there are more results, the response includes a marker that you can use to request the next set of results. You can also request initialized documents.

This operation returns paginated results.

Creating a Request

describeFolderContents

Arguments

:: Text
dfcsFolderId

-> DescribeFolderContents

Creates a value of DescribeFolderContents with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dfcsInclude - The contents to include. Specify INITIALIZED to include initialized documents.

dfcsAuthenticationToken - Amazon WorkDocs authentication token. This field should not be set when using administrative API actions, as in accessing the API using AWS credentials.

dfcsSort - The sorting criteria.

dfcsMarker - The marker for the next set of results. This marker was received from a previous call.

dfcsLimit - The maximum number of items to return with this call.

dfcsType - The type of items.

dfcsOrder - The order for the contents of the folder.

dfcsFolderId - The ID of the folder.

data DescribeFolderContents

See: describeFolderContents smart constructor.

Request Lenses

dfcsInclude :: Lens' DescribeFolderContents (Maybe Text)

The contents to include. Specify INITIALIZED to include initialized documents.

dfcsAuthenticationToken :: Lens' DescribeFolderContents (Maybe Text)

Amazon WorkDocs authentication token. This field should not be set when using administrative API actions, as in accessing the API using AWS credentials.

dfcsSort :: Lens' DescribeFolderContents (Maybe ResourceSortType)

The sorting criteria.

dfcsMarker :: Lens' DescribeFolderContents (Maybe Text)

The marker for the next set of results. This marker was received from a previous call.

dfcsLimit :: Lens' DescribeFolderContents (Maybe Natural)

The maximum number of items to return with this call.

dfcsType :: Lens' DescribeFolderContents (Maybe FolderContentType)

The type of items.

dfcsOrder :: Lens' DescribeFolderContents (Maybe OrderType)

The order for the contents of the folder.

dfcsFolderId :: Lens' DescribeFolderContents Text

The ID of the folder.

Destructuring the Response

describeFolderContentsResponse

Arguments

:: Int
dfcrsResponseStatus

-> DescribeFolderContentsResponse

Creates a value of DescribeFolderContentsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dfcrsFolders - The subfolders in the specified folder.

dfcrsDocuments - The documents in the specified folder.

dfcrsMarker - The marker to use when requesting the next set of results. If there are no additional results, the string is empty.

dfcrsResponseStatus - -- | The response status code.

data DescribeFolderContentsResponse

See: describeFolderContentsResponse smart constructor.

Response Lenses

dfcrsFolders :: Lens' DescribeFolderContentsResponse [FolderMetadata]

The subfolders in the specified folder.

dfcrsDocuments :: Lens' DescribeFolderContentsResponse [DocumentMetadata]

The documents in the specified folder.

dfcrsMarker :: Lens' DescribeFolderContentsResponse (Maybe Text)

The marker to use when requesting the next set of results. If there are no additional results, the string is empty.

dfcrsResponseStatus :: Lens' DescribeFolderContentsResponse Int

- | The response status code.