Network.AWS.WorkDocs.DescribeNotificationSubscriptions

Description

Lists the specified notification subscriptions.

Creating a Request

describeNotificationSubscriptions

Arguments

:: Text
dOrganizationId

-> DescribeNotificationSubscriptions

Creates a value of DescribeNotificationSubscriptions with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dMarker - The marker for the next set of results. (You received this marker from a previous call.)

dLimit - The maximum number of items to return with this call.

dOrganizationId - The ID of the organization.

data DescribeNotificationSubscriptions

See: describeNotificationSubscriptions smart constructor.

Request Lenses

dMarker :: Lens' DescribeNotificationSubscriptions (Maybe Text)

The marker for the next set of results. (You received this marker from a previous call.)

dLimit :: Lens' DescribeNotificationSubscriptions (Maybe Natural)

The maximum number of items to return with this call.

dOrganizationId :: Lens' DescribeNotificationSubscriptions Text

The ID of the organization.

Destructuring the Response

describeNotificationSubscriptionsResponse

Arguments

:: Int
dnsrsResponseStatus

-> DescribeNotificationSubscriptionsResponse

Creates a value of DescribeNotificationSubscriptionsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dnsrsMarker - The marker to use when requesting the next set of results. If there are no additional results, the string is empty.

dnsrsSubscriptions - The subscriptions.

dnsrsResponseStatus - -- | The response status code.

data DescribeNotificationSubscriptionsResponse

See: describeNotificationSubscriptionsResponse smart constructor.

Response Lenses

dnsrsMarker :: Lens' DescribeNotificationSubscriptionsResponse (Maybe Text)

The marker to use when requesting the next set of results. If there are no additional results, the string is empty.

dnsrsSubscriptions :: Lens' DescribeNotificationSubscriptionsResponse [Subscription]

The subscriptions.

dnsrsResponseStatus :: Lens' DescribeNotificationSubscriptionsResponse Int

- | The response status code.