Network.AWS.WorkDocs.DescribeRootFolders

Description

Describes the current user's special folders; the RootFolder and the RecyleBin . RootFolder is the root of user's files and folders and RecyleBin is the root of recycled items. This is not a valid action for SigV4 (administrative API) clients.

Creating a Request

describeRootFolders

Arguments

:: Text
drfAuthenticationToken

-> DescribeRootFolders

Creates a value of DescribeRootFolders with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

drfMarker - The marker for the next set of results. (You received this marker from a previous call.)

drfLimit - The maximum number of items to return.

drfAuthenticationToken - Amazon WorkDocs authentication token. This field should not be set when using administrative API actions, as in accessing the API using AWS credentials.

data DescribeRootFolders

See: describeRootFolders smart constructor.

Request Lenses

drfMarker :: Lens' DescribeRootFolders (Maybe Text)

The marker for the next set of results. (You received this marker from a previous call.)

drfLimit :: Lens' DescribeRootFolders (Maybe Natural)

The maximum number of items to return.

drfAuthenticationToken :: Lens' DescribeRootFolders Text

Amazon WorkDocs authentication token. This field should not be set when using administrative API actions, as in accessing the API using AWS credentials.

Destructuring the Response

describeRootFoldersResponse

Arguments

:: Int
drfrsResponseStatus

-> DescribeRootFoldersResponse

Creates a value of DescribeRootFoldersResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

drfrsFolders - The user's special folders.

drfrsMarker - The marker for the next set of results.

drfrsResponseStatus - -- | The response status code.

data DescribeRootFoldersResponse

See: describeRootFoldersResponse smart constructor.

Response Lenses

drfrsFolders :: Lens' DescribeRootFoldersResponse [FolderMetadata]

The user's special folders.

drfrsMarker :: Lens' DescribeRootFoldersResponse (Maybe Text)

The marker for the next set of results.

drfrsResponseStatus :: Lens' DescribeRootFoldersResponse Int

- | The response status code.