Network.AWS.WorkDocs.DescribeUsers

Description

Describes the specified users. You can describe all users or filter the results (for example, by status or organization).

By default, Amazon WorkDocs returns the first 24 active or pending users. If there are more results, the response includes a marker that you can use to request the next set of results.

This operation returns paginated results.

Creating a Request

describeUsers :: DescribeUsers

Creates a value of DescribeUsers with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

duInclude - The state of the users. Specify ALL to include inactive users.

duUserIds - The IDs of the users.

duAuthenticationToken - Amazon WorkDocs authentication token. This field should not be set when using administrative API actions, as in accessing the API using AWS credentials.

duSort - The sorting criteria.

duMarker - The marker for the next set of results. (You received this marker from a previous call.)

duQuery - A query to filter users by user name.

duLimit - The maximum number of items to return.

duOrder - The order for the results.

duOrganizationId - The ID of the organization.

duFields - A comma-separated list of values. Specify STORAGE_METADATA to include the user storage quota and utilization information.

data DescribeUsers

See: describeUsers smart constructor.

Request Lenses

duInclude :: Lens' DescribeUsers (Maybe UserFilterType)

The state of the users. Specify ALL to include inactive users.

duUserIds :: Lens' DescribeUsers (Maybe Text)

The IDs of the users.

duAuthenticationToken :: Lens' DescribeUsers (Maybe Text)

Amazon WorkDocs authentication token. This field should not be set when using administrative API actions, as in accessing the API using AWS credentials.

duSort :: Lens' DescribeUsers (Maybe UserSortType)

The sorting criteria.

duMarker :: Lens' DescribeUsers (Maybe Text)

The marker for the next set of results. (You received this marker from a previous call.)

duQuery :: Lens' DescribeUsers (Maybe Text)

A query to filter users by user name.

duLimit :: Lens' DescribeUsers (Maybe Natural)

The maximum number of items to return.

duOrder :: Lens' DescribeUsers (Maybe OrderType)

The order for the results.

duOrganizationId :: Lens' DescribeUsers (Maybe Text)

The ID of the organization.

duFields :: Lens' DescribeUsers (Maybe Text)

A comma-separated list of values. Specify STORAGE_METADATA to include the user storage quota and utilization information.

Destructuring the Response

describeUsersResponse

Arguments

:: Int
dursResponseStatus

-> DescribeUsersResponse

Creates a value of DescribeUsersResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dursUsers - The users.

dursTotalNumberOfUsers - The total number of users included in the results.

dursMarker - The marker to use when requesting the next set of results. If there are no additional results, the string is empty.

dursResponseStatus - -- | The response status code.

data DescribeUsersResponse

See: describeUsersResponse smart constructor.

Response Lenses

dursUsers :: Lens' DescribeUsersResponse [User]

The users.

dursTotalNumberOfUsers :: Lens' DescribeUsersResponse (Maybe Integer)

The total number of users included in the results.

dursMarker :: Lens' DescribeUsersResponse (Maybe Text)

The marker to use when requesting the next set of results. If there are no additional results, the string is empty.

dursResponseStatus :: Lens' DescribeUsersResponse Int

- | The response status code.