Network.AWS.WorkDocs.GetDocument

Description

Retrieves details of a document.

Creating a Request

getDocument

Arguments

:: Text
gdDocumentId

-> GetDocument

Creates a value of GetDocument with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gdAuthenticationToken - Amazon WorkDocs authentication token. This field should not be set when using administrative API actions, as in accessing the API using AWS credentials.

gdIncludeCustomMetadata - Set this to TRUE to include custom metadata in the response.

gdDocumentId - The ID of the document.

data GetDocument

See: getDocument smart constructor.

Request Lenses

gdAuthenticationToken :: Lens' GetDocument (Maybe Text)

Amazon WorkDocs authentication token. This field should not be set when using administrative API actions, as in accessing the API using AWS credentials.

gdIncludeCustomMetadata :: Lens' GetDocument (Maybe Bool)

Set this to TRUE to include custom metadata in the response.

gdDocumentId :: Lens' GetDocument Text

The ID of the document.

Destructuring the Response

getDocumentResponse

Arguments

:: Int
gdrsResponseStatus

-> GetDocumentResponse

Creates a value of GetDocumentResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gdrsCustomMetadata - The custom metadata on the document.

gdrsMetadata - The metadata details of the document.

gdrsResponseStatus - -- | The response status code.

data GetDocumentResponse

See: getDocumentResponse smart constructor.

Response Lenses

gdrsCustomMetadata :: Lens' GetDocumentResponse (HashMap Text Text)

The custom metadata on the document.

gdrsMetadata :: Lens' GetDocumentResponse (Maybe DocumentMetadata)

The metadata details of the document.

gdrsResponseStatus :: Lens' GetDocumentResponse Int

- | The response status code.