Network.AWS.WorkDocs.GetDocumentPath

Description

Retrieves the path information (the hierarchy from the root folder) for the requested document.

By default, Amazon WorkDocs returns a maximum of 100 levels upwards from the requested document and only includes the IDs of the parent folders in the path. You can limit the maximum number of levels. You can also request the names of the parent folders.

Creating a Request

getDocumentPath

Arguments

:: Text
gdpDocumentId

-> GetDocumentPath

Creates a value of GetDocumentPath with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gdpAuthenticationToken - Amazon WorkDocs authentication token. This field should not be set when using administrative API actions, as in accessing the API using AWS credentials.

gdpMarker - This value is not supported.

gdpLimit - The maximum number of levels in the hierarchy to return.

gdpFields - A comma-separated list of values. Specify NAME to include the names of the parent folders.

gdpDocumentId - The ID of the document.

data GetDocumentPath

See: getDocumentPath smart constructor.

Request Lenses

gdpAuthenticationToken :: Lens' GetDocumentPath (Maybe Text)

Amazon WorkDocs authentication token. This field should not be set when using administrative API actions, as in accessing the API using AWS credentials.

gdpMarker :: Lens' GetDocumentPath (Maybe Text)

This value is not supported.

gdpLimit :: Lens' GetDocumentPath (Maybe Natural)

The maximum number of levels in the hierarchy to return.

gdpFields :: Lens' GetDocumentPath (Maybe Text)

A comma-separated list of values. Specify NAME to include the names of the parent folders.

gdpDocumentId :: Lens' GetDocumentPath Text

The ID of the document.

Destructuring the Response

getDocumentPathResponse

Arguments

:: Int
gdprsResponseStatus

-> GetDocumentPathResponse

Creates a value of GetDocumentPathResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gdprsPath - The path information.

gdprsResponseStatus - -- | The response status code.

data GetDocumentPathResponse

See: getDocumentPathResponse smart constructor.

Response Lenses

gdprsPath :: Lens' GetDocumentPathResponse (Maybe ResourcePath)

The path information.

gdprsResponseStatus :: Lens' GetDocumentPathResponse Int

- | The response status code.