Network.AWS.WorkDocs.GetDocumentVersion

Description

Retrieves version metadata for the specified document.

Creating a Request

getDocumentVersion

Arguments

:: Text
gdvDocumentId

-> Text
gdvVersionId

-> GetDocumentVersion

Creates a value of GetDocumentVersion with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gdvAuthenticationToken - Amazon WorkDocs authentication token. This field should not be set when using administrative API actions, as in accessing the API using AWS credentials.

gdvIncludeCustomMetadata - Set this to TRUE to include custom metadata in the response.

gdvFields - A comma-separated list of values. Specify SOURCE to include a URL for the source document.

gdvDocumentId - The ID of the document.

gdvVersionId - The version ID of the document.

data GetDocumentVersion

See: getDocumentVersion smart constructor.

Request Lenses

gdvAuthenticationToken :: Lens' GetDocumentVersion (Maybe Text)

Amazon WorkDocs authentication token. This field should not be set when using administrative API actions, as in accessing the API using AWS credentials.

gdvIncludeCustomMetadata :: Lens' GetDocumentVersion (Maybe Bool)

Set this to TRUE to include custom metadata in the response.

gdvFields :: Lens' GetDocumentVersion (Maybe Text)

A comma-separated list of values. Specify SOURCE to include a URL for the source document.

gdvDocumentId :: Lens' GetDocumentVersion Text

The ID of the document.

gdvVersionId :: Lens' GetDocumentVersion Text

The version ID of the document.

Destructuring the Response

getDocumentVersionResponse

Arguments

:: Int
gdvrsResponseStatus

-> GetDocumentVersionResponse

Creates a value of GetDocumentVersionResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gdvrsCustomMetadata - The custom metadata on the document version.

gdvrsMetadata - The version metadata.

gdvrsResponseStatus - -- | The response status code.

data GetDocumentVersionResponse

See: getDocumentVersionResponse smart constructor.

Response Lenses

gdvrsCustomMetadata :: Lens' GetDocumentVersionResponse (HashMap Text Text)

The custom metadata on the document version.

gdvrsMetadata :: Lens' GetDocumentVersionResponse (Maybe DocumentVersionMetadata)

The version metadata.

gdvrsResponseStatus :: Lens' GetDocumentVersionResponse Int

- | The response status code.