Network.AWS.WorkDocs.GetFolderPath

Description

Retrieves the path information (the hierarchy from the root folder) for the specified folder.

By default, Amazon WorkDocs returns a maximum of 100 levels upwards from the requested folder and only includes the IDs of the parent folders in the path. You can limit the maximum number of levels. You can also request the parent folder names.

Creating a Request

getFolderPath

Arguments

:: Text
gfpFolderId

-> GetFolderPath

Creates a value of GetFolderPath with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gfpAuthenticationToken - Amazon WorkDocs authentication token. This field should not be set when using administrative API actions, as in accessing the API using AWS credentials.

gfpMarker - This value is not supported.

gfpLimit - The maximum number of levels in the hierarchy to return.

gfpFields - A comma-separated list of values. Specify NAME to include the names of the parent folders.

gfpFolderId - The ID of the folder.

data GetFolderPath

See: getFolderPath smart constructor.

Request Lenses

gfpAuthenticationToken :: Lens' GetFolderPath (Maybe Text)

Amazon WorkDocs authentication token. This field should not be set when using administrative API actions, as in accessing the API using AWS credentials.

gfpMarker :: Lens' GetFolderPath (Maybe Text)

This value is not supported.

gfpLimit :: Lens' GetFolderPath (Maybe Natural)

The maximum number of levels in the hierarchy to return.

gfpFields :: Lens' GetFolderPath (Maybe Text)

A comma-separated list of values. Specify NAME to include the names of the parent folders.

gfpFolderId :: Lens' GetFolderPath Text

The ID of the folder.

Destructuring the Response

getFolderPathResponse

Arguments

:: Int
gfprsResponseStatus

-> GetFolderPathResponse

Creates a value of GetFolderPathResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gfprsPath - The path information.

gfprsResponseStatus - -- | The response status code.

data GetFolderPathResponse

See: getFolderPathResponse smart constructor.

Response Lenses

gfprsPath :: Lens' GetFolderPathResponse (Maybe ResourcePath)

The path information.

gfprsResponseStatus :: Lens' GetFolderPathResponse Int

- | The response status code.