Network.AWS.WorkDocs.InitiateDocumentVersionUpload

Description

Creates a new document object and version object.

The client specifies the parent folder ID and name of the document to upload. The ID is optionally specified when creating a new version of an existing document. This is the first step to upload a document. Next, upload the document to the URL returned from the call, and then call UpdateDocumentVersion .

To cancel the document upload, call AbortDocumentVersionUpload .

Creating a Request

initiateDocumentVersionUpload

Arguments

:: Text
idvuParentFolderId

-> InitiateDocumentVersionUpload

Creates a value of InitiateDocumentVersionUpload with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

idvuDocumentSizeInBytes - The size of the document, in bytes.

idvuContentCreatedTimestamp - The time stamp when the content of the document was originally created.

idvuAuthenticationToken - Amazon WorkDocs authentication token. This field should not be set when using administrative API actions, as in accessing the API using AWS credentials.

idvuName - The name of the document.

idvuId - The ID of the document.

idvuContentModifiedTimestamp - The time stamp when the content of the document was modified.

idvuContentType - The content type of the document.

idvuParentFolderId - The ID of the parent folder.

data InitiateDocumentVersionUpload

See: initiateDocumentVersionUpload smart constructor.

Request Lenses

idvuDocumentSizeInBytes :: Lens' InitiateDocumentVersionUpload (Maybe Integer)

The size of the document, in bytes.

idvuContentCreatedTimestamp :: Lens' InitiateDocumentVersionUpload (Maybe UTCTime)

The time stamp when the content of the document was originally created.

idvuAuthenticationToken :: Lens' InitiateDocumentVersionUpload (Maybe Text)

Amazon WorkDocs authentication token. This field should not be set when using administrative API actions, as in accessing the API using AWS credentials.

idvuName :: Lens' InitiateDocumentVersionUpload (Maybe Text)

The name of the document.

idvuId :: Lens' InitiateDocumentVersionUpload (Maybe Text)

The ID of the document.

idvuContentModifiedTimestamp :: Lens' InitiateDocumentVersionUpload (Maybe UTCTime)

The time stamp when the content of the document was modified.

idvuContentType :: Lens' InitiateDocumentVersionUpload (Maybe Text)

The content type of the document.

idvuParentFolderId :: Lens' InitiateDocumentVersionUpload Text

The ID of the parent folder.

Destructuring the Response

initiateDocumentVersionUploadResponse

Arguments

:: Int
idvursResponseStatus

-> InitiateDocumentVersionUploadResponse

Creates a value of InitiateDocumentVersionUploadResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

idvursMetadata - The document metadata.

idvursUploadMetadata - The upload metadata.

idvursResponseStatus - -- | The response status code.

data InitiateDocumentVersionUploadResponse

See: initiateDocumentVersionUploadResponse smart constructor.

Response Lenses

idvursMetadata :: Lens' InitiateDocumentVersionUploadResponse (Maybe DocumentMetadata)

The document metadata.

idvursUploadMetadata :: Lens' InitiateDocumentVersionUploadResponse (Maybe UploadMetadata)

The upload metadata.

idvursResponseStatus :: Lens' InitiateDocumentVersionUploadResponse Int

- | The response status code.