Network.AWS.WorkDocs.RemoveResourcePermission

Description

Removes the permission for the specified principal from the specified resource.

Creating a Request

removeResourcePermission

Arguments

:: Text
rrpResourceId

-> Text
rrpPrincipalId

-> RemoveResourcePermission

Creates a value of RemoveResourcePermission with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

rrpPrincipalType - The principal type of the resource.

rrpAuthenticationToken - Amazon WorkDocs authentication token. This field should not be set when using administrative API actions, as in accessing the API using AWS credentials.

rrpResourceId - The ID of the resource.

rrpPrincipalId - The principal ID of the resource.

data RemoveResourcePermission

See: removeResourcePermission smart constructor.

Request Lenses

rrpPrincipalType :: Lens' RemoveResourcePermission (Maybe PrincipalType)

The principal type of the resource.

rrpAuthenticationToken :: Lens' RemoveResourcePermission (Maybe Text)

Amazon WorkDocs authentication token. This field should not be set when using administrative API actions, as in accessing the API using AWS credentials.

rrpResourceId :: Lens' RemoveResourcePermission Text

The ID of the resource.

rrpPrincipalId :: Lens' RemoveResourcePermission Text

The principal ID of the resource.

Destructuring the Response

removeResourcePermissionResponse :: RemoveResourcePermissionResponse

Creates a value of RemoveResourcePermissionResponse with the minimum fields required to make a request.

data RemoveResourcePermissionResponse

See: removeResourcePermissionResponse smart constructor.