Network.AWS.WorkDocs.UpdateDocument

Description

Updates the specified attributes of a document. The user must have access to both the document and its parent folder, if applicable.

Creating a Request

updateDocument

Arguments

:: Text
udDocumentId

-> UpdateDocument

Creates a value of UpdateDocument with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

udParentFolderId - The ID of the parent folder.

udAuthenticationToken - Amazon WorkDocs authentication token. This field should not be set when using administrative API actions, as in accessing the API using AWS credentials.

udName - The name of the document.

udResourceState - The resource state of the document. Note that only ACTIVE and RECYCLED are supported.

udDocumentId - The ID of the document.

data UpdateDocument

See: updateDocument smart constructor.

Request Lenses

udParentFolderId :: Lens' UpdateDocument (Maybe Text)

The ID of the parent folder.

udAuthenticationToken :: Lens' UpdateDocument (Maybe Text)

Amazon WorkDocs authentication token. This field should not be set when using administrative API actions, as in accessing the API using AWS credentials.

udName :: Lens' UpdateDocument (Maybe Text)

The name of the document.

udResourceState :: Lens' UpdateDocument (Maybe ResourceStateType)

The resource state of the document. Note that only ACTIVE and RECYCLED are supported.

udDocumentId :: Lens' UpdateDocument Text

The ID of the document.

Destructuring the Response

updateDocumentResponse :: UpdateDocumentResponse

Creates a value of UpdateDocumentResponse with the minimum fields required to make a request.

data UpdateDocumentResponse

See: updateDocumentResponse smart constructor.