Network.AWS.WorkDocs.UpdateDocumentVersion

Description

Changes the status of the document version to ACTIVE.

Amazon WorkDocs also sets its document container to ACTIVE. This is the last step in a document upload, after the client uploads the document to an S3-presigned URL returned by InitiateDocumentVersionUpload .

Creating a Request

updateDocumentVersion

Arguments

:: Text
udvDocumentId

-> Text
udvVersionId

-> UpdateDocumentVersion

Creates a value of UpdateDocumentVersion with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

udvAuthenticationToken - Amazon WorkDocs authentication token. This field should not be set when using administrative API actions, as in accessing the API using AWS credentials.

udvVersionStatus - The status of the version.

udvDocumentId - The ID of the document.

udvVersionId - The version ID of the document.

data UpdateDocumentVersion

See: updateDocumentVersion smart constructor.

Request Lenses

udvAuthenticationToken :: Lens' UpdateDocumentVersion (Maybe Text)

Amazon WorkDocs authentication token. This field should not be set when using administrative API actions, as in accessing the API using AWS credentials.

udvVersionStatus :: Lens' UpdateDocumentVersion (Maybe DocumentVersionStatus)

The status of the version.

udvDocumentId :: Lens' UpdateDocumentVersion Text

The ID of the document.

udvVersionId :: Lens' UpdateDocumentVersion Text

The version ID of the document.

Destructuring the Response

updateDocumentVersionResponse :: UpdateDocumentVersionResponse

Creates a value of UpdateDocumentVersionResponse with the minimum fields required to make a request.

data UpdateDocumentVersionResponse

See: updateDocumentVersionResponse smart constructor.